Suany Arrieta: Realizo la pagina de donaciones y su diseño ( En proceso ). __Paula Marin: Realizo el map para mostrar los datos almacenados en la base de datos( En proceso ) y correccion de errores realacionados con el CRUD.

+++ FrontEnd/Clinica_Dental/src/Service/Servicios/DeleteServicios.jsx
@@ -1,9 +1,17 @@
 async function DeleteServicios(id) {
+
+    
     try {
-        const response = await fetch(`http://localhost:3001/servicios/${id}`, {
+    const token = localStorage.getItem('token');
+
+    if (!token) {
+      throw new Error('Token no encontrado en localStorage');
+    }
+        const response = await fetch(`http://127.0.0.1:8000/api/servicios/${id}`, {
             method: 'DELETE',
             headers: {
-                'Content-Type': 'application/json'
+                'Content-Type': 'application/json',
+                'Authorization': `Bearer ${token}`,
             }
         });
