+++ FrontEnd/Clinica_Dental/src/Service/Usuarios/DeleteUsers.jsx
@@ -1,10 +1,18 @@
 // Su funcion es eliminar datos servidor
 async function DeleteUsers(id) {
+
     try {
-        const response = await fetch(`http://127.0.0.1:8000/api/token/${id}`, {
+    const token = localStorage.getItem('token');
+
+    if (!token) {
+      throw new Error('Token no encontrado en localStorage');
+    }
+    
+        const response = await fetch(`http://127.0.0.1:8000/api/users/${id}`, {
             method: 'DELETE',
             headers: {
-                'Content-Type': 'application/json'
+                'Content-Type': 'application/json',
+                'Authorization': `Bearer ${token}`,
             }
         });
