+++ FrontEnd/Clinica_Dental/src/Service/testimonios/DeleteTestimonios.jsx
@@ -1,9 +1,17 @@
 async function DeleteTestimonials(id) {
+
+    
     try {
+    const token = localStorage.getItem('token');
+
+    if (!token) {
+      throw new Error('Token no encontrado en localStorage');
+    }
         const response = await fetch(`http://127.0.0.1:8000/api/testimonios/${id}`, {
             method: 'DELETE',
             headers: {
-                'Content-Type': 'application/json'
+                'Content-Type': 'application/json',
+                'Authorization': `Bearer ${token}`,
             }
         });
